requirements.txt
pymongo==4.1.1
discord.py==2.4.0
beautifulsoup4==4.11.1
gTTS==2.3.2
PyNaCl==1.5.0
requests==2.31.0
dnspython==2.7.0
openai==0.28
Django==5.1.6
mcstatus==12.0.1
asyncpg==0.30.0
pytz==2025.2
paramiko==4.0.0